+++ services/kms/model_wrapping_key.go
@@ -384,8 +384,7 @@ type WrappingKey struct {
 	// REQUIRED
 	CreatedAt WrappingKeyGetCreatedAtAttributeType `json:"createdAt" required:"true"`
 	// A user chosen description to distinguish multiple wrapping keys.
-	// REQUIRED
-	Description WrappingKeyGetDescriptionAttributeType `json:"description" required:"true"`
+	Description WrappingKeyGetDescriptionAttributeType `json:"description,omitempty"`
 	// The display name to distinguish multiple wrapping keys.
 	// REQUIRED
 	DisplayName WrappingKeyGetDisplayNameAttributeType `json:"displayName" required:"true"`
@@ -415,12 +414,11 @@ type _WrappingKey WrappingKey
 // This constructor will assign default values to properties that have it defined,
 // and makes sure properties required by API are set, but the set of arguments
 // will change when the set of required properties is changed
-func NewWrappingKey(accessScope WrappingKeyGetAccessScopeArgType, algorithm WrappingKeyGetAlgorithmArgType, createdAt WrappingKeyGetCreatedAtArgType, description WrappingKeyGetDescriptionArgType, displayName WrappingKeyGetDisplayNameArgType, expiresAt WrappingKeyGetExpiresAtArgType, id WrappingKeyGetIdArgType, keyRingId WrappingKeyGetKeyRingIdArgType, protection WrappingKeyGetProtectionArgType, publicKey WrappingKeyGetPublicKeyArgType, purpose WrappingKeyGetPurposeArgType, state WrappingKeyGetStateArgType) *WrappingKey {
+func NewWrappingKey(accessScope WrappingKeyGetAccessScopeArgType, algorithm WrappingKeyGetAlgorithmArgType, createdAt WrappingKeyGetCreatedAtArgType, displayName WrappingKeyGetDisplayNameArgType, expiresAt WrappingKeyGetExpiresAtArgType, id WrappingKeyGetIdArgType, keyRingId WrappingKeyGetKeyRingIdArgType, protection WrappingKeyGetProtectionArgType, publicKey WrappingKeyGetPublicKeyArgType, purpose WrappingKeyGetPurposeArgType, state WrappingKeyGetStateArgType) *WrappingKey {
 	this := WrappingKey{}
 	setWrappingKeyGetAccessScopeAttributeType(&this.AccessScope, accessScope)
 	setWrappingKeyGetAlgorithmAttributeType(&this.Algorithm, algorithm)
 	setWrappingKeyGetCreatedAtAttributeType(&this.CreatedAt, createdAt)
-	setWrappingKeyGetDescriptionAttributeType(&this.Description, description)
 	setWrappingKeyGetDisplayNameAttributeType(&this.DisplayName, displayName)
 	setWrappingKeyGetExpiresAtAttributeType(&this.ExpiresAt, expiresAt)
 	setWrappingKeyGetIdAttributeType(&this.Id, id)
@@ -493,19 +491,25 @@ func (o *WrappingKey) SetCreatedAt(v WrappingKeyGetCreatedAtRetType) {
 	setWrappingKeyGetCreatedAtAttributeType(&o.CreatedAt, v)
 }
 
-// GetDescription returns the Description field value
-func (o *WrappingKey) GetDescription() (ret WrappingKeyGetDescriptionRetType) {
-	ret, _ = o.GetDescriptionOk()
-	return ret
+// GetDescription returns the Description field value if set, zero value otherwise.
+func (o *WrappingKey) GetDescription() (res WrappingKeyGetDescriptionRetType) {
+	res, _ = o.GetDescriptionOk()
+	return
 }
 
-// GetDescriptionOk returns a tuple with the Description field value
+// GetDescriptionOk returns a tuple with the Description field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *WrappingKey) GetDescriptionOk() (ret WrappingKeyGetDescriptionRetType, ok bool) {
 	return getWrappingKeyGetDescriptionAttributeTypeOk(o.Description)
 }
 
-// SetDescription sets field value
+// HasDescription returns a boolean if a field has been set.
+func (o *WrappingKey) HasDescription() bool {
+	_, ok := o.GetDescriptionOk()
+	return ok
+}
+
+// SetDescription gets a reference to the given string and assigns it to the Description field.
 func (o *WrappingKey) SetDescription(v WrappingKeyGetDescriptionRetType) {
 	setWrappingKeyGetDescriptionAttributeType(&o.Description, v)
 }