+++ CHANGELOG.md
@@ -113,6 +113,10 @@
   - [v0.1.0](services/logs/CHANGELOG.md#v010)
     - **New:** API for logs service
 - `kms`: 
+  - [v1.3.0](services/kms/CHANGELOG.md#v130)
+    - Set field `Description` to optional in struct `Key`, `KeyRing` and `WrappingKey`
+      - **Breaking change:** Remove `description` parameter from `NewKey`, `NewKeyRing` and `NewWrappingKey` constructors
+      - **Feature:** Add method `HasDescription`in `Key`, `KeyRing` and `WrappingKey` models
   - [v1.2.1](services/kms/CHANGELOG.md#v121)
     - **Bugfix:** Correctly handle file closing for file uploads
     - Bump STACKIT SDK core module from `v0.20.1` to `v0.21.0`